Remove references to doomed Parse implementation
Parse is shutting down, so remove all the references to it. The files
that used to live under "src/cloud" have been moved to "src/server".
`npm run start` continues to work.

In another commit I'll be adding a replacement for Parse, which is the
ability to run from a static set of files (which means being able to run
from GitHub Pages or similar).